added support for compiled multi includes
