+++ src/Marten.Testing/Documents/Issue.cs
@@ -15,6 +15,8 @@ namespace Marten.Testing.Documents
 
         public string Title { get; set; }
 
+        public int Number { get; set; }
+
         public Guid? AssigneeId { get; set; }
 
         public Guid? ReporterId { get; set; }
