+++ src/Marten.Testing/Marten.Testing.csproj
@@ -180,6 +180,7 @@
     <Compile Include="Schema\DbObjectsTests.cs" />
     <Compile Include="Schema\SchemaDiff_Generation_Tests.cs" />
     <Compile Include="Schema\TableNameTests.cs" />
+    <Compile Include="Services\Includes\end_to_end_query_with_compiled_include_Tests.cs" />
     <Compile Include="Services\SelectTransformerTests.cs" />
     <Compile Include="Linq\select_transformations_Tests.cs" />
     <Compile Include="Linq\SingleFieldSelectorTests.cs" />
