+++ src/Marten.Testing/Services/Includes/end_to_end_query_with_include_Tests.cs
@@ -104,41 +104,6 @@ namespace Marten.Testing.Services.Includes
         // ENDSAMPLE
 
         [Fact]
-        public void simple_compiled_include_for_a_single_document()
-        {
-            var user = new User();
-            var issue = new Issue { AssigneeId = user.Id, Title = "Garage Door is busted" };
-
-            theSession.Store<object>(user, issue);
-            theSession.SaveChanges();
-
-            using (var query = theStore.QuerySession())
-            {
-                var issueQuery = new IssueByTitleIncludingUser {Title = issue.Title};
-                var issue2 = query.Query(issueQuery);
-
-                issueQuery.Included.ShouldNotBeNull();
-                issueQuery.Included.Id.ShouldBe(user.Id);
-
-                issue2.ShouldNotBeNull();
-            }
-        }
-
-        public class IssueByTitleIncludingUser : ICompiledQuery<Issue>
-        {
-            public string Title { get; set; }
-            public User Included { get; private set; } = new User();
-            public JoinType JoinType { get; set; } = JoinType.Inner;
-
-            public Expression<Func<IQueryable<Issue>, Issue>> QueryIs()
-            {
-                return query => query
-                    .Include<Issue, IssueByTitleIncludingUser>(x => x.AssigneeId, x => x.Included, JoinType)
-                    .Single(x => x.Title == Title);
-            }
-        }
-
-        [Fact]
         public void include_with_containment_where_for_a_single_document()
         {
             var user = new User();
